# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Tasker",
        "description": "This is a sample Petstore server.  You can find\nout more about Swagger at\n[http://swagger.io](http://swagger.io) or on\n[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/comments/{id}/attachments": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Upload attachment to a comment",
                "operationId": "App\\Http\\Controllers\\AttachmentController::comment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of comment",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "attachment": {
                                        "description": "file to upload",
                                        "type": "string",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return success"
                    }
                }
            }
        },
...
        "/attachment/{id}": {
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Remove an attachment",
                "operationId": "App\\Http\\Controllers\\AttachmentController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of attachment",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return success"
                    }
                }
            }
        },
...
        "/tasks/{slug}/bids": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create a new bid",
                "operationId": "App\\Http\\Controllers\\BidController::store",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of task",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "comment": {
                                        "description": "Bid's body",
                                        "type": "string",
                                        "example": "This is a bid comment"
                                    },
                                    "price": {
                                        "description": "Bid's price",
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Createa a comment successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/tasks/{slug}/comments": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create a new comment",
                "operationId": "App\\Http\\Controllers\\CommentController::store",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of task",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "description": "Comment's body",
                                        "type": "string",
                                        "example": "This is a comments"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Createa a comment successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/comments/{id}/replies": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Create a new comment reply",
                "operationId": "App\\Http\\Controllers\\CommentController::reply",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of comment",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create data format",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "description": "Comment's body",
                                        "type": "string",
                                        "example": "This is a comments"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Createa a comment successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    }
}